Bug 56057 - JMeter does not create ApacheJMeterTemporaryRootCA.crt when starting a recording Add a checks for generation of CRT before copying it Bugzilla Id: 56057
git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1624740 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 494e8926105f1537650c74142f927faa1e757948

+++ src/jorphan/org/apache/jorphan/exec/KeyToolUtils.java
@@ -251,7 +251,11 @@ public class KeyToolUtils {
         // Export the Root CA for Firefox/Chrome/IE
         KeyToolUtils.keytool("-exportcert", keystore, password, ROOTCA_ALIAS, null, null, "-rfc", "-file", ROOT_CACERT_CRT);
         // Copy for Opera
-        FileUtils.copyFile(caCert_crt, caCert_usr);
+        if(caCert_crt.exists() && caCert_crt.canRead()) {
+            FileUtils.copyFile(caCert_crt, caCert_usr);            
+        } else {
+            log.warn("Failed creating "+caCert_crt.getAbsolutePath()+", check 'keytool' utility in path is available and points to a JDK >= 7");
+        }
     }
 
     /**
